Move subtitles into its own component
